Redirect properly after a user logs in
If a user is not logged in and they try to do something that requires
them to be logged in, they are redirected to the root path and asked to
log in. After they log in, they are redirected back to the original page
that required them to be logged in.

If a user is on a page and they login, they will get redirected back to
that page as the session[:referrer_url] will be nil. Therefore the
`request.env['omniauth.origin']` will be used.